+++ src/FAI3_backend/src/inference_providers/mod.rs
@@ -1,0 +1,19 @@
+pub mod traits;
+pub mod novita;
+pub mod together;
+pub mod nebius;
+pub mod none;
+pub mod lib;
+
+// Re-export the trait and providers
+pub use traits::InferenceProvider;
+pub use novita::NovitaProvider;
+pub use together::TogetherAIProvider;
+pub use nebius::NebiusProvider;
+pub use none::NoneProvider;
+pub use lib::{
+    HuggingFaceRequestParameters,
+    HuggingFaceRequest,
+    HUGGING_FACE_ENDPOINT,
+    HUGGING_FACE_INFERENCE_PROVIDER_URL,
+};
